% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CppLHS}
\alias{CppLHS}
\title{This is the internal Cpp function used to run the metropolis hasting algorithm if use.cpp = T. 
In general, it shouldn't be used as a stand alone function, because some preprocessing is done in R}
\arguments{
\item{xA}{matrix of data - must be numeric (factors are converted to numeric in R)}

\item{cost}{cost vector (0 if no cost)}

\item{strata}{matrix of continuous strata}

\item{include}{matrix of included data}

\item{idx}{integer vector of rows from which sampling is allowed}

\item{factors}{boolean factor flag}

\item{i_fact}{indices of factors in xA}

\item{nsample}{number of samples}

\item{cost_mode}{bool cost flag}

\item{iter}{number of iterations}

\item{wCont}{continuous weight}

\item{wFact}{factor weights}

\item{wCorr}{correlation weights}

\item{etaMat}{eta matrix - either all 1, or user input}

\item{temperature}{initial temperature}

\item{tdecrease}{temperature decrease every length_cycle iterations}

\item{length_cycle}{number of iterations between temperature decrease}
}
\value{
list with sampled data, indices, objective values, cost value, and final continuous weights for each sample
}
\description{
This is the internal Cpp function used to run the metropolis hasting algorithm if use.cpp = T. 
In general, it shouldn't be used as a stand alone function, because some preprocessing is done in R
}
