% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clhs-internal.R
\name{.lhs_obj}
\alias{.lhs_obj}
\title{Objective function for cLHS}
\usage{
.lhs_obj(
  size,
  data_continuous_sampled,
  data_factor_sampled,
  continuous_strata,
  weights = list(numeric = 1, factor = 1, corelation = 1),
  cor_mat,
  factor_obj,
  eta = 1
)
}
\description{
Objective function for cLHS
}
\author{
Pierre Roudier
}
\keyword{internal}
